pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openapi-mcp"
version = "0.1.0"
description = "Convert OpenAPI specifications into Model Context Protocol servers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "OpenAPI-MCP Contributors"}
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp>=0.1.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "requests>=2.0.0",
    "wheel>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.0.254",
    "mypy>=1.0.0",
    "types-requests>=2.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
openapi-mcp = "openapi_mcp.cli:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.11.0",
]
